+++ forui/test/src/widgets/dialog/dialog_golden_test.dart
@@ -38,8 +38,13 @@ void main() {
                   'This action cannot be undone. This will permanently delete your account and remove your data from our servers.',
                 ),
                 actions: [
-                  FButton(child: const Text('Continue'), onPress: () {}),
-                  FButton(style: FButtonStyle.outline, child: const Text('Cancel'), onPress: () {}),
+                  FButton(onPress: () {}, intrinsicWidth: false, child: const Text('Continue')),
+                  FButton(
+                    style: FButtonStyle.outline,
+                    intrinsicWidth: false,
+                    onPress: () {},
+                    child: const Text('Cancel'),
+                  ),
                 ],
               ),
             ),